requirements.txt
bcrypt==4.0.1
click==8.1.3
Flask==2.2.2
Flask-Bcrypt==1.0.1
Flask-Cors==3.0.10
Flask-Login==0.6.2
gunicorn==20.1.0
importlib-metadata==5.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
peewee==3.15.3
psycopg-binary==3.1.4
psycopg2==2.9.5
python-dotenv==0.21.0
six==1.16.0
Werkzeug==2.2.2
zipp==3.10.0
